feat: 다이얼로그 컴포넌트 구성, 위젯 레이어 이관작업
