+++ src/widgets/comment-dialog/ui/index.ts
@@ -1,0 +1,2 @@
+export { CommentAddDialog } from "./CommentAddDialog"
+export { CommentUpdateDialog } from "./CommentUpdateDialog"
