+++ src/widgets/post-dialog/ui/index.ts
@@ -1,0 +1,3 @@
+export { PostAddDialog } from "./PostAddDialog"
+export { PostDetailDialog } from "./PostDetailDialog"
+export { PostUpdateDialog } from "./PostUpdateDialog"
