+++ src/widgets/user-dialog/ui/index.ts
@@ -1,0 +1,1 @@
+export { UserInfoDialog } from "./UserInfoDialog"
